completed payment gateway 2.0

+++ app/api/products/payProduct/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { doc, getDoc, updateDoc } from "firebase/firestore";
+import { updateStock } from "../../products/putProduct/updateStock/updateStock";
 import { db } from "@/firebase/config";
 const mercadopago = require("mercadopago");
 
@@ -25,46 +26,51 @@ export const POST = async (req: NextRequest) => {
       const { status, additional_info, external_reference } =
         await paymentInfo.json();
 
+      let userId = external_reference;
+
       let products = await additional_info?.items;
 
-      let arrayProducts = [];
-
       if (status === "approved") {
-        await fetch(
-          "https://discount-dash-53vw-git-develop-ufcampos.vercel.app/api/cart/delete",
-          {
-            method: "DELETE",
-            body: JSON.stringify({
-              userId: external_reference,
-            }),
-          }
-        );
+        let arrayProducts = [];
+        if (userId.charAt(0) === "-") {
+          userId = userId.slice(1);
+
+          await fetch(
+            "https://discount-dash-53vw-git-develop-ufcampos.vercel.app/api/cart/delete",
+            {
+              method: "DELETE",
+              body: JSON.stringify({
+                userId,
+              }),
+            }
+          );
+        }
         for (let item of products) {
-          const reference = doc(db, "products", item?.id);
-          const productDB = await getDoc(reference);
-          const info = await productDB.data();
-          if (info) {
-            info.id = item?.id;
-            info.quantity = Number(item?.quantity);
-          }
+          const response = await fetch(
+            `https://discount-dash-53vw-git-develop-ufcampos.vercel.app/api/products/${item?.id}`
+          );
+          const info = await response.json();
+          if (info) info.quantity = Number(item?.quantity);
+
           await arrayProducts.push(info);
 
+          const reference = doc(db, "products", item?.id);
           updateDoc(reference, {
             stock: info?.stock - Number(item.quantity),
           });
         }
-      }
 
-      await fetch(
-        "https://discount-dash-53vw-git-develop-ufcampos.vercel.app/api/shopOrder/post",
-        {
-          method: "POST",
-          body: JSON.stringify({
-            userId: external_reference,
-            arrayProducts,
-          }),
-        }
-      );
+        await fetch(
+          "https://discount-dash-53vw-git-develop-ufcampos.vercel.app/api/shopOrder/post",
+          {
+            method: "POST",
+            body: JSON.stringify({
+              userId,
+              arrayProducts,
+            }),
+          }
+        );
+      }
     }
     return NextResponse.json({ response: "Data received" });
   } catch (error) {
